clean commons.py
