+++ qcip_tools/chemistry_files/helpers.py
@@ -3,7 +3,7 @@ import argparse
 
 from qcip_tools import chemistry_files
 # Note: we need to explicitly import all modules, so that they are inspected!
-from qcip_tools.chemistry_files import gamess, gaussian, dalton, xyz, chemistry_datafile, pdb  # noqa
+from qcip_tools.chemistry_files import gamess, gaussian, dalton, xyz, chemistry_datafile, pdb, csv_tensor, stda  # noqa
 
 
 class ProbablyNotAChemistryFile(Exception):
