+++ qcip_tools/scripts/excitations.py
@@ -7,9 +7,9 @@ import argparse
 import sys
 import math
 
+import qcip_tools.scripts
 from qcip_tools import derivatives_exci, derivatives_e
 from qcip_tools.chemistry_files import helpers, PropertyNotPresent
-from qcip_tools.scripts import commons
 
 __version__ = '0.1'
 __author__ = 'Pierre Beaujean'
@@ -50,18 +50,18 @@ def main():
     args = parser.parse_args()
 
     if not args.infile.has_property('excitations'):
-        commons.exit_failure('cannot find excitations ({})'.format(args.infile.file_type))
+        qcip_tools.scripts.exit_failure('cannot find excitations ({})'.format(args.infile.file_type))
 
     try:
         excitations = args.infile.property('excitations')
     except PropertyNotPresent as e:
-        return commons.exit_failure('cannot find {} ({})'.format(e, args.infile.file_type))
+        return qcip_tools.scripts.exit_failure('cannot find {} ({})'.format(e, args.infile.file_type))
 
     if '!F' not in excitations:
-        return commons.exit_failure('no transition dipoles (!F)')
+        return qcip_tools.scripts.exit_failure('no transition dipoles (!F)')
 
     if '!' not in excitations and '#' not in excitations:
-        return commons.exit_failure('no excitation energies (# or !)')
+        return qcip_tools.scripts.exit_failure('no excitation energies (# or !)')
 
     has_S2 = '_<S2>' in excitations
     if has_S2:
